localservicebookingnew/backend_django/requirements.txt
Django==4.2.29
djangorestframework==3.17.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-allauth==65.15.0
PyMySQL==1.1.2
python-decouple==3.8
qrcode==8.2
Pillow==12.1.1
cryptography==46.0.5
requests==2.32.5
